Refactored code to implement step pattern
